requirements.txt
requests
apprise
pyyaml
